refactor get_case_transactions to also fetch for an existing list of transactions
