+++ corehq/form_processor/backends/sql/processor.py
@@ -24,6 +24,7 @@ from corehq.form_processor.models import (
     XFormInstanceSQL, XFormAttachmentSQL, CaseTransaction,
     CommCareCaseSQL, FormEditRebuild, Attachment, XFormOperationSQL)
 from corehq.form_processor.utils import convert_xform_to_json, extract_meta_instance_id, extract_meta_user_id
+from corehq.form_processor.utils.sql import get_case_transactions_by_case_id
 from couchforms.const import ATTACHMENT_NAME
 from dimagi.utils.couch import acquire_lock, release_lock
 import six
@@ -304,7 +305,7 @@ class FormProcessorSQL(object):
 
     @staticmethod
     def _rebuild_case_from_transactions(case, detail, updated_xforms=None):
-        transactions = get_case_transactions(case.case_id, updated_xforms=updated_xforms)
+        transactions = get_case_transactions_by_case_id(case.case_id, updated_xforms=updated_xforms)
         strategy = SqlCaseUpdateStrategy(case)
 
         rebuild_transaction = CaseTransaction.rebuild_transaction(case, detail)
@@ -342,43 +343,3 @@ class FormProcessorSQL(object):
     @staticmethod
     def case_exists(case_id):
         return CaseAccessorSQL.case_exists(case_id)
-
-
-def get_case_transactions(case_id, updated_xforms=None):
-    """
-    This fetches all the transactions required to rebuild the case along
-    with all the forms for those transactions.
-
-    For any forms that have been updated it replaces the old form
-    with the new one.
-
-    :param case_id: ID of case to rebuild
-    :param updated_xforms: list of forms that have been changed.
-    :return: list of ``CaseTransaction`` objects with their associated forms attached.
-    """
-    transactions = CaseAccessorSQL.get_transactions_for_case_rebuild(case_id)
-    form_ids = {tx.form_id for tx in transactions}
-    updated_xforms_map = {
-        xform.form_id: xform for xform in updated_xforms if not xform.is_deprecated
-    } if updated_xforms else {}
-
-    form_ids_to_fetch = list(form_ids - set(updated_xforms_map.keys()))
-    xform_map = {form.form_id: form for form in FormAccessorSQL.get_forms_with_attachments_meta(form_ids_to_fetch)}
-
-    def get_form(form_id):
-        if form_id in updated_xforms_map:
-            return updated_xforms_map[form_id]
-
-        try:
-            return xform_map[form_id]
-        except KeyError:
-            raise XFormNotFound
-
-    for transaction in transactions:
-        if transaction.form_id:
-            try:
-                transaction.cached_form = get_form(transaction.form_id)
-            except XFormNotFound:
-                logging.error('Form not found during rebuild: %s', transaction.form_id)
-
-    return transactions
